style: make workspace clippy -D warnings green (clippy 1.94 drift)
Mechanical, behavior-neutral clippy fixes across the workspace so
`cargo clippy --workspace --all-targets -- -D warnings` passes on the
clippy 1.94 toolchain. No feature work, no API changes, no behavior
changes.

- Idiom rewrites: match_result_ok, single_match, for_kv_map,
  unnecessary_cast, needless_borrow(s), manual_repeat_n, useless_conversion,
  manual_c_str_literals, manual_range_contains, redundant_closure,
  useless_vec, get_first, collapsible_match, vec_init_then_push,
  expect_fun_call, needless_return, cloned_ref_to_slice_refs.
- items_after_test_module: byte-identical function relocations
  (hayro-jbig2, hayro-jpeg2000, pdf-interpret/cid, pdf-manip/pdfa_cleanup,
  xfa-cli, pdf-render).
- field_reassign_with_default: test setup rewritten to struct-literal form.
- Scoped #[allow] with rationale where the lint is a literal-test-data or
  contract-assertion false positive (approx_constant 3.14,
  assertions_on_constants const floors, cloned_ref_to_slice_refs where the
  page is also moved into the comparison slice).

Fork edits (pdfluent-lopdf, hayro-*) are confined to test code / function
relocations only.

+++ crates/pdf-manip/examples/debug_cff_glyph_widths.rs
@@ -68,7 +68,7 @@ fn main() {
             let gid = cff_parser::GlyphId(i);
             match table.glyph_width(gid) {
                 Some(w) => {
-                    let scaled = (w as f64 * scale as f64).round() as i64;
+                    let scaled = (w as f64 * scale).round() as i64;
                     let cid = table.glyph_cid(gid).unwrap_or(i);
                     widths.push((cid, scaled));
                 }
